de-structure users and vendor
