+++ src/Features/Admin/Users/AddUser/index.js
@@ -1,10 +1,9 @@
 import React, { useState } from 'react';
 
 import { useMutation } from 'react-query';
-import { useSelector } from 'react-redux';
 
 import CommonGridBasedForm from '../../../../components/CommonGridBasedForm';
-import { SELECT, TEXT_FIELD } from '../../../../components/CommonGridBasedForm/FieldTypes';
+import { TEXT_FIELD } from '../../../../components/CommonGridBasedForm/FieldTypes';
 import { emailRegex } from '../../../../redux/ActionTypes';
 import { contactRegex, GetHeader } from '../../../../scripts/constants';
 import { validateOnSubmit, fieldChangeHandler } from '../../../../util/CommonGridBasedFormUtils';
@@ -12,17 +11,8 @@ import { createUser } from '../../Common Requests/mutation';
 
 const AddUser = () => {
   const { headers } = GetHeader();
-  const { token } = useSelector((state) => {
-    const {
-      authReducer: {
-        accessToken: { token },
-      },
-    } = state;
-    return {
-      token,
-    };
-  });
-  const AddUser = useMutation(createUser, {
+
+  const { isLoading, mutateAsync, isSuccess } = useMutation(createUser, {
     onSuccess: () => {
       const resetFields = fields.map((field) => {
         return {
@@ -34,19 +24,6 @@ const AddUser = () => {
     },
   });
   const [fields, setFields] = useState([
-    {
-      type: SELECT,
-      label: 'Role',
-      values: ['user', 'vendor'],
-      value: [],
-      name: 'role',
-      errorMessage: '',
-
-      onChange: ({ target: { value } }, index) => {
-        const updatedFields = fieldChangeHandler(fields, value, index);
-        setFields(updatedFields);
-      },
-    },
     {
       type: TEXT_FIELD,
       textFieldType: 'text',
@@ -126,13 +103,12 @@ const AddUser = () => {
 
     if (isValid) {
       const userData = {};
+      userData['role'] = 'user';
       fields.map(({ name, value }) => {
-        if (name !== 'building') {
-          userData[name] = value;
-        }
+        userData[name] = value;
       });
 
-      AddUser.mutateAsync({ headers, userData });
+      mutateAsync({ headers, userData });
     }
   };
 
@@ -142,11 +118,11 @@ const AddUser = () => {
       name: 'save',
       minWidth: '100%',
       clickHandler: saveHandler,
-      isLoading: AddUser.isLoading,
+      isLoading,
     },
   ];
 
-  return <CommonGridBasedForm buttons={buttons} fields={fields} heading="Add User" onSaveSuccess={AddUser.isSuccess} />;
+  return <CommonGridBasedForm buttons={buttons} fields={fields} heading="Add User" onSaveSuccess={isSuccess} />;
 };
 
 export default AddUser;
